pyproject.toml
[project]
name = "poormans-whatsapp"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "black>=26.1.0",
    "celery>=5.6.2",
    "fastapi>=0.134.0",
    "isort>=8.0.0",
    "openai>=2.24.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "redis>=7.2.1",
    "ruff>=0.15.4",
    "sqlalchemy[asyncio]>=2.0.47",
    "whatsapp-python>=4.3.0",
]

[dependency-groups]
dev = [
    "commitizen>=4.13.9",
]
